Refactor MovementSystem and CharacterBuilder classes
The MovementSystem script has been simplified for cleaner code. This includes removing unnecessary conditional statements and updating the process of setting target locations. In the CharacterBuilder class, the EntityManager initialization has been optimized for reusability. Also, adaptations were made to the character building process in correspondence to these changes. Adjustment was also made to TransformSyncSystem and the Wolf Testing.asset file for configuration.